refactor(rel): extract custom aggregate domain (#1162)
